Logger testing
